[M680x0] Pass the correct CPU name to LLVM
For some reasons, we only passed numerical CPU name (e.g. 68010) to
LLVM, which was not recognized by it. This commit not only fixed this
but also provides several options for user to specify CPU in the driver.

This fixes #14

In addition, it turned out that 68060 CPU is not implemented in LLVM
backend, so remove it from Clang as well.